src/agent/requirements.txt
azure-ai-projects>=1.0.0
azure-ai-agents>=1.0.0b1
azure-identity>=1.19.0
python-dotenv>=1.0.1
